Fix registration application list
In b02a676bebd2cb4ebb24473eec678136424bbd19 the registration application list
was changed to use the `validator_time` of the local user instead of the
`published` time of the person. The `validator_time` is no longer present in
the Lemmy API, which broke the current implementation. This reverts the logic
back to using just the person's `published` timestamp.

fixes #37